+++ Microsoft.Azure.Cosmos/src/CosmosClientOptions.cs
@@ -8,7 +8,6 @@ namespace Microsoft.Azure.Cosmos
     using System.Collections.ObjectModel;
     using System.Data.Common;
     using System.Linq;
-    using System.Runtime.ConstrainedExecution;
     using Microsoft.Azure.Cosmos.Fluent;
     using Microsoft.Azure.Documents;
     using Microsoft.Azure.Documents.Client;
@@ -17,6 +16,18 @@ namespace Microsoft.Azure.Cosmos
     /// <summary>
     /// Defines all the configurable options that the CosmosClient requires.
     /// </summary>
+    /// <example>
+    /// An example on how to configure the serialization option to ignore null values
+    /// CosmosClientOptions clientOptions = new CosmosClientOptions()
+    /// {
+    ///     SerializerOptions = new CosmosSerializationOptions(){
+    ///         IgnoreNullValues = true
+    ///     },
+    ///     ConnectionMode = ConnectionMode.Gateway,
+    /// };
+    /// 
+    /// CosmosClient client = new CosmosClient("endpoint", "key", clientOptions);
+    /// </example>
     public class CosmosClientOptions
     {
         /// <summary>
@@ -39,9 +50,19 @@ namespace Microsoft.Azure.Cosmos
         /// </summary>
         private static readonly CosmosSerializer propertiesSerializer = new CosmosJsonSerializerWrapper(new CosmosJsonDotNetSerializer());
 
-        private readonly Collection<RequestHandler> customHandlers;
+        private readonly string currentEnvironmentInformation;
 
         private int gatewayModeMaxConnectionLimit;
+        private string applicationName;
+        private CosmosSerializationOptions serializerOptions;
+        private CosmosSerializer serializer;
+
+        private ConnectionMode connectionMode;
+        private Protocol connectionProtocol;
+        private TimeSpan? idleTcpConnectionTimeout;
+        private TimeSpan? openTcpConnectionTimeout;
+        private int? maxRequestsPerTcpConnection;
+        private int? maxTcpConnectionsPerEndpoint;
 
         /// <summary>
         /// Creates a new CosmosClientOptions
@@ -49,12 +70,15 @@ namespace Microsoft.Azure.Cosmos
         public CosmosClientOptions()
         {
             this.UserAgentContainer = new UserAgentContainer();
+            EnvironmentInformation environmentInformation = new EnvironmentInformation();
+            this.currentEnvironmentInformation = environmentInformation.ToString();
+            this.UserAgentContainer.Suffix = this.currentEnvironmentInformation;
             this.GatewayModeMaxConnectionLimit = ConnectionPolicy.Default.MaxConnectionLimit;
             this.RequestTimeout = ConnectionPolicy.Default.RequestTimeout;
             this.ConnectionMode = CosmosClientOptions.DefaultConnectionMode;
             this.ConnectionProtocol = CosmosClientOptions.DefaultProtocol;
             this.ApiType = CosmosClientOptions.DefaultApiType;
-            this.customHandlers = new Collection<RequestHandler>();
+            this.CustomHandlers = new Collection<RequestHandler>();
         }
 
         /// <summary>
@@ -65,8 +89,12 @@ namespace Microsoft.Azure.Cosmos
         /// </remarks>
         public string ApplicationName
         {
-            get => this.UserAgentContainer.Suffix;
-            set => this.UserAgentContainer.Suffix = value;
+            get => this.applicationName;
+            set
+            {
+                this.UserAgentContainer.Suffix = this.currentEnvironmentInformation + EnvironmentInformation.Delimiter + value;
+                this.applicationName = value;
+            }
         }
 
         /// <summary>
@@ -123,10 +151,7 @@ namespace Microsoft.Azure.Cosmos
         /// </summary>
         /// <seealso cref="CosmosClientBuilder.AddCustomHandlers(RequestHandler[])"/>
         [JsonConverter(typeof(ClientOptionJsonConverter))]
-        public Collection<RequestHandler> CustomHandlers
-        {
-            get => this.customHandlers;
-        }
+        public Collection<RequestHandler> CustomHandlers { get; }
 
         /// <summary>
         /// Get or set the connection mode used by the client when connecting to the Azure Cosmos DB service.
@@ -137,9 +162,17 @@ namespace Microsoft.Azure.Cosmos
         /// <remarks>
         /// For more information, see <see href="https://docs.microsoft.com/azure/documentdb/documentdb-performance-tips#direct-connection">Connection policy: Use direct connection mode</see>.
         /// </remarks>
-        /// <seealso cref="CosmosClientBuilder.WithConnectionModeDirect"/>
+        /// <seealso cref="CosmosClientBuilder.WithConnectionModeDirect()"/>
         /// <seealso cref="CosmosClientBuilder.WithConnectionModeGateway(int?)"/>
-        public ConnectionMode ConnectionMode { get; set; }
+        public ConnectionMode ConnectionMode
+        {
+            get => this.connectionMode;
+            set
+            {
+                this.ValidateDirectTCPSettings();
+                this.connectionMode = value;
+            }
+        }
 
         /// <summary>
         /// This can be used to weaken the database account consistency level for read operations.
@@ -148,7 +181,7 @@ namespace Microsoft.Azure.Cosmos
         public ConsistencyLevel? ConsistencyLevel { get; set; }
 
         /// <summary>
-        /// Get ot set the number of times client should retry on rate throttled requests.
+        /// Get or set the number of times client should retry on rate throttled requests.
         /// </summary>
         /// <seealso cref="CosmosClientBuilder.WithThrottlingRetryOptions(TimeSpan, int)"/>
         public int? MaxRetryAttemptsOnRateLimitedRequests { get; set; }
@@ -163,6 +196,112 @@ namespace Microsoft.Azure.Cosmos
         public TimeSpan? MaxRetryWaitTimeOnRateLimitedRequests { get; set; }
 
         /// <summary>
+        /// (Direct/TCP) Controls the amount of idle time after which unused connections are closed.
+        /// </summary>
+        /// <value>
+        /// By default, idle connections are kept open indefinitely. Value must be greater than or equal to 10 minutes. Recommended values are between 20 minutes and 24 hours.
+        /// </value>
+        /// <remarks>
+        /// Mainly useful for sparse infrequent access to a large database account.
+        /// </remarks>
+        public TimeSpan? IdleTcpConnectionTimeout
+        {
+            get => this.idleTcpConnectionTimeout;
+            set
+            {
+                this.idleTcpConnectionTimeout = value;
+                this.ValidateDirectTCPSettings();                
+            }
+        }      
+
+        /// <summary>
+        /// (Direct/TCP) Controls the amount of time allowed for trying to establish a connection.
+        /// </summary>
+        /// <value>
+        /// The default timeout is 5 seconds. Recommended values are greater than or equal to 5 seconds.
+        /// </value>
+        /// <remarks>
+        /// When the time elapses, the attempt is cancelled and an error is returned. Longer timeouts will delay retries and failures.
+        /// </remarks>
+        public TimeSpan? OpenTcpConnectionTimeout
+        {
+            get => this.openTcpConnectionTimeout;
+            set
+            {
+                this.openTcpConnectionTimeout = value;
+                this.ValidateDirectTCPSettings();                
+            }
+        }
+
+        /// <summary>
+        /// (Direct/TCP) Controls the number of requests allowed simultaneously over a single TCP connection. When more requests are in flight simultaneously, the direct/TCP client will open additional connections.
+        /// </summary>
+        /// <value>
+        /// The default settings allow 30 simultaneous requests per connection.
+        /// Do not set this value lower than 4 requests per connection or higher than 50-100 requests per connection.       
+        /// The former can lead to a large number of connections to be created. 
+        /// The latter can lead to head of line blocking, high latency and timeouts.
+        /// </value>
+        /// <remarks>
+        /// Applications with a very high degree of parallelism per connection, with large requests or responses, or with very tight latency requirements might get better performance with 8-16 requests per connection.
+        /// </remarks>
+        public int? MaxRequestsPerTcpConnection
+        {
+            get => this.maxRequestsPerTcpConnection;
+            set
+            {
+                this.maxRequestsPerTcpConnection = value;
+                this.ValidateDirectTCPSettings();                
+            }
+        }
+
+        /// <summary>
+        /// (Direct/TCP) Controls the maximum number of TCP connections that may be opened to each Cosmos DB back-end.
+        /// Together with MaxRequestsPerTcpConnection, this setting limits the number of requests that are simultaneously sent to a single Cosmos DB back-end(MaxRequestsPerTcpConnection x MaxTcpConnectionPerEndpoint).
+        /// </summary>
+        /// <value>
+        /// The default value is 65,535. Value must be greater than or equal to 16.
+        /// </value>
+        public int? MaxTcpConnectionsPerEndpoint
+        {           
+            get => this.maxTcpConnectionsPerEndpoint;
+            set
+            {
+                this.maxTcpConnectionsPerEndpoint = value;
+                this.ValidateDirectTCPSettings();                
+            }
+        }
+
+        /// <summary>
+        /// Get to set optional serializer options.
+        /// </summary>
+        /// <example>
+        /// An example on how to configure the serialization option to ignore null values
+        /// CosmosClientOptions clientOptions = new CosmosClientOptions()
+        /// {
+        ///     SerializerOptions = new CosmosSerializationOptions(){
+        ///         IgnoreNullValues = true
+        ///     }
+        /// };
+        /// 
+        /// CosmosClient client = new CosmosClient("endpoint", "key", clientOptions);
+        /// </example>
+        public CosmosSerializationOptions SerializerOptions
+        {
+            get => this.serializerOptions;
+            set
+            {
+                if (this.Serializer != null)
+                {
+                    throw new ArgumentException(
+                        $"{nameof(this.SerializerOptions)} is not compatible with {nameof(this.Serializer)}. Only one can be set.  ");
+                }
+
+                this.serializerOptions = value;
+            }
+        }
+
+        /// <summary>
         /// Get to set an optional JSON serializer. The client will use it to serialize or de-serialize user's cosmos request/responses.
         /// SDK owned types such as DatabaseProperties and ContainerProperties will always use the SDK default serializer.
         /// </summary>
@@ -173,10 +312,7 @@ namespace Microsoft.Azure.Cosmos
         /// <example>
         /// // An example on how to configure the serializer to ignore null values
         /// CosmosSerializer ignoreNullSerializer = new CosmosJsonDotNetSerializer(
-        ///         new JsonSerializerSettings()
-        ///         {
-        ///             NullValueHandling = NullValueHandling.Ignore
-        ///         });
+        ///             NullValueHandling = NullValueHandling.Ignore);
         ///         
         /// CosmosClientOptions clientOptions = new CosmosClientOptions()
         /// {
@@ -186,7 +322,20 @@ namespace Microsoft.Azure.Cosmos
         /// CosmosClient client = new CosmosClient("endpoint", "key", clientOptions);
         /// </example>
         [JsonConverter(typeof(ClientOptionJsonConverter))]
-        public CosmosSerializer Serializer { get; set; }
+        public CosmosSerializer Serializer
+        {
+            get => this.serializer;
+            set
+            {
+                if (this.SerializerOptions != null)
+                {
+                    throw new ArgumentException(
+                        $"{nameof(this.Serializer)} is not compatible with {nameof(this.SerializerOptions)}. Only one can be set.  ");
+                }
+
+                this.serializer = value;
+            }
+        }
 
         /// <summary>
         /// A JSON serializer used by the CosmosClient to serialize or de-serialize cosmos request/responses.
@@ -195,12 +344,6 @@ namespace Microsoft.Azure.Cosmos
         /// </summary>
         [JsonConverter(typeof(ClientOptionJsonConverter))]
         internal CosmosSerializer PropertiesSerializer => CosmosClientOptions.propertiesSerializer;
-
-        /// <summary>
-        /// Gets the user json serializer with the CosmosJsonSerializerWrapper or the default
-        /// </summary>
-        [JsonIgnore]
-        internal CosmosSerializer CosmosSerializerWithWrapperOrDefault => this.Serializer == null ? this.PropertiesSerializer : new CosmosJsonSerializerWrapper(this.Serializer);
 
         /// <summary>
         /// Gets or sets the connection protocol when connecting to the Azure Cosmos service.
@@ -213,7 +356,15 @@ namespace Microsoft.Azure.Cosmos
         /// Gateway mode only supports HTTPS.
         /// For more information, see <see href="https://docs.microsoft.com/azure/documentdb/documentdb-performance-tips#use-tcp">Connection policy: Use the TCP protocol</see>.
         /// </remarks>
-        internal Protocol ConnectionProtocol { get; set; }
+        internal Protocol ConnectionProtocol
+        {
+            get => this.connectionProtocol;
+            set
+            {
+                this.ValidateDirectTCPSettings();
+                this.connectionProtocol = value;
+            }
+        }
 
         internal UserAgentContainer UserAgentContainer { get; private set; }
 
@@ -300,6 +451,22 @@ namespace Microsoft.Azure.Cosmos
         /// </summary>
         internal bool? EnableCpuMonitor { get; set; }
 
+        /// <summary>
+        /// Gets the user json serializer with the CosmosJsonSerializerWrapper or the default
+        /// </summary>
+        internal CosmosSerializer GetCosmosSerializerWithWrapperOrDefault()
+        {
+            if (this.SerializerOptions != null)
+            {
+                CosmosJsonDotNetSerializer cosmosJsonDotNetSerializer = new CosmosJsonDotNetSerializer(this.SerializerOptions);
+                return new CosmosJsonSerializerWrapper(cosmosJsonDotNetSerializer);
+            }
+            else
+            {
+                return this.Serializer == null ? this.PropertiesSerializer : new CosmosJsonSerializerWrapper(this.Serializer);
+            }
+        }
+
         internal CosmosClientOptions Clone()
         {
             CosmosClientOptions cloneConfiguration = (CosmosClientOptions)this.MemberwiseClone();
@@ -308,6 +475,7 @@ namespace Microsoft.Azure.Cosmos
 
         internal ConnectionPolicy GetConnectionPolicy()
         {
+            this.ValidateDirectTCPSettings();
             ConnectionPolicy connectionPolicy = new ConnectionPolicy()
             {
                 MaxConnectionLimit = this.GatewayModeMaxConnectionLimit,
@@ -316,6 +484,10 @@ namespace Microsoft.Azure.Cosmos
                 ConnectionProtocol = this.ConnectionProtocol,
                 UserAgentContainer = this.UserAgentContainer,
                 UseMultipleWriteLocations = true,
+                IdleTcpConnectionTimeout = this.IdleTcpConnectionTimeout,
+                OpenTcpConnectionTimeout = this.OpenTcpConnectionTimeout,
+                MaxRequestsPerTcpConnection = this.MaxRequestsPerTcpConnection,
+                MaxTcpConnectionsPerEndpoint = this.MaxTcpConnectionsPerEndpoint
             };
 
             if (this.ApplicationRegion != null)
@@ -412,6 +584,35 @@ namespace Microsoft.Azure.Cosmos
             }
 
             throw new ArgumentException("The connection string is missing a required property: " + keyName);
+        }
+
+        private void ValidateDirectTCPSettings()
+        {
+            string settingName = string.Empty;
+            if (!(this.ConnectionMode == ConnectionMode.Direct && this.ConnectionProtocol == Protocol.Tcp))
+            {
+                if (this.IdleTcpConnectionTimeout.HasValue)
+                {
+                    settingName = nameof(this.IdleTcpConnectionTimeout);
+                }
+                else if (this.OpenTcpConnectionTimeout.HasValue)
+                {
+                    settingName = nameof(this.OpenTcpConnectionTimeout);
+                }
+                else if (this.MaxRequestsPerTcpConnection.HasValue)
+                {
+                    settingName = nameof(this.MaxRequestsPerTcpConnection);
+                }
+                else if (this.MaxTcpConnectionsPerEndpoint.HasValue)
+                {
+                    settingName = nameof(this.MaxTcpConnectionsPerEndpoint);
+                }
+            }
+
+            if (!string.IsNullOrEmpty(settingName))
+            {
+                throw new ArgumentException($"{settingName} requires {nameof(this.ConnectionMode)} to be set to {nameof(ConnectionMode.Direct)} and {nameof(this.ConnectionProtocol)} to be set to {nameof(Protocol.Tcp)}");
+            }            
         }
 
         /// <summary>