+++ Microsoft.Azure.Cosmos/src/CosmosElements/CosmosElementEqualityComparer.cs
@@ -8,7 +8,14 @@ namespace Microsoft.Azure.Cosmos.CosmosElements
     using System.Collections.Generic;
     using System.Linq;
 
-    internal sealed class CosmosElementEqualityComparer : IEqualityComparer<CosmosElement>
+#if INTERNAL
+#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
+#pragma warning disable SA1600 // Elements should be documented
+    public
+#else
+    internal
+#endif
+    sealed class CosmosElementEqualityComparer : IEqualityComparer<CosmosElement>
     {
         public static CosmosElementEqualityComparer Value = new CosmosElementEqualityComparer();
 
@@ -192,4 +199,8 @@ namespace Microsoft.Azure.Cosmos.CosmosElements
             return 0;
         }
     }
+#if INTERNAL
+#pragma warning restore SA1600 // Elements should be documented
+#pragma warning restore CS1591 // Missing XML comment for publicly visible type or member
+#endif
 }