+++ Microsoft.Azure.Cosmos/src/RequestOptions/QueryRequestOptions.cs
@@ -150,12 +150,14 @@ namespace Microsoft.Azure.Cosmos
         /// </remarks>
         internal string SessionToken { get; set; }
 
-        internal CosmosSerializationOptions CosmosSerializationOptions { get; set; }
+        internal CosmosSerializationFormatOptions CosmosSerializationOptions { get; set; }
 
         /// <summary>
         /// Gets or sets the flag that enables skip take across partitions.
         /// </summary>
         internal bool EnableCrossPartitionSkipTake { get; set; }
+
+        internal bool EnableGroupBy { get; set; }
 
         /// <summary>
         /// Fill the CosmosRequestMessage headers with the set properties
@@ -232,6 +234,7 @@ namespace Microsoft.Azure.Cosmos
                 PartitionKey = this.PartitionKey,
                 CosmosSerializationOptions = this.CosmosSerializationOptions,
                 EnableCrossPartitionSkipTake = this.EnableCrossPartitionSkipTake,
+                EnableGroupBy = this.EnableGroupBy,
                 Properties = this.Properties,
                 IsEffectivePartitionKeyRouting = this.IsEffectivePartitionKeyRouting,
                 PopulateQueryMetrics = this.PopulateQueryMetrics
@@ -252,6 +255,7 @@ namespace Microsoft.Azure.Cosmos
                 MaxBufferedItemCount = this.MaxBufferedItemCount.HasValue ? this.MaxBufferedItemCount.Value : 0,
                 CosmosSerializationOptions = this.CosmosSerializationOptions,
                 Properties = this.Properties,
+                EnableGroupBy = this.EnableGroupBy
             };
         }