+++ Microsoft.Azure.Cosmos/src/Resource/QueryResponses/QueryResponse.cs
@@ -76,7 +76,8 @@ namespace Microsoft.Azure.Cosmos
 
         internal IReadOnlyDictionary<string, QueryMetrics> queryMetrics { get; set; }
 
-        internal virtual CosmosSerializationOptions CosmosSerializationOptions { get; set; }
+        internal virtual CosmosSerializationFormatOptions CosmosSerializationOptions { get; set; }
+
 
         internal bool GetHasMoreResults()
         {
@@ -144,64 +145,42 @@ namespace Microsoft.Azure.Cosmos
     {
         private readonly IEnumerable<CosmosElement> cosmosElements;
         private readonly CosmosSerializer jsonSerializer;
-        private readonly CosmosSerializationOptions serializationOptions;
+        private readonly CosmosSerializationFormatOptions serializationOptions;
         private IEnumerable<T> resources;
 
-        /// <summary>
-        /// Create a <see cref="QueryResponse{T}"/>
-        /// </summary>
         private QueryResponse(
+            HttpStatusCode httpStatusCode,
             IEnumerable<CosmosElement> cosmosElements,
             CosmosQueryResponseMessageHeaders responseMessageHeaders,
             CosmosDiagnostic cosmosDiagnostic,
             CosmosSerializer jsonSerializer,
-            CosmosSerializationOptions serializationOptions)
+            CosmosSerializationFormatOptions serializationOptions)
         {
             this.cosmosElements = cosmosElements;
             this.QueryHeaders = responseMessageHeaders;
             this.cosmosDiagnostic = cosmosDiagnostic;
             this.jsonSerializer = jsonSerializer;
             this.serializationOptions = serializationOptions;
+            this.StatusCode = httpStatusCode;
         }
 
-        /// <summary>
-        /// Gets the continuation token
-        /// </summary>
         public override string ContinuationToken => this.Headers.ContinuationToken;
 
-        /// <summary>
-        /// Gets the request charge for this request from the Azure Cosmos DB service.
-        /// </summary>
-        /// <value>
-        /// The request charge measured in request units.
-        /// </value>
         public override double RequestCharge => this.Headers.RequestCharge;
 
-        /// <summary>
-        /// The headers of the response
-        /// </summary>
         public override Headers Headers => this.QueryHeaders;
 
-        /// <summary>
-        /// The number of items in the stream.
-        /// </summary>
+        public override HttpStatusCode StatusCode { get; }
+
         public override int Count => this.cosmosElements?.Count() ?? 0;
 
         internal CosmosQueryResponseMessageHeaders QueryHeaders { get; }
 
-        /// <summary>
-        /// Get the enumerators to iterate through the results
-        /// </summary>
-        /// <returns>An enumerator of the response objects</returns>
         public override IEnumerator<T> GetEnumerator()
         {
             return this.Resource.GetEnumerator();
         }
 
-        /// <summary>
-        /// Get the enumerators to iterate through the results
-        /// </summary>
-        /// <returns>An enumerator of the response objects</returns>
         public override IEnumerable<T> Resource
         {
             get
@@ -229,6 +208,7 @@ namespace Microsoft.Azure.Cosmos
             {
                 cosmosQueryResponse.EnsureSuccessStatusCode();
                 queryResponse = new QueryResponse<TInput>(
+                    httpStatusCode: cosmosQueryResponse.StatusCode,
                     cosmosElements: cosmosQueryResponse.CosmosElements,
                     responseMessageHeaders: cosmosQueryResponse.QueryHeaders,
                     cosmosDiagnostic: cosmosQueryResponse.cosmosDiagnostic,